requirements.txt
pyTelegramBotAPI
gspread
oauth2client
python-dotenv
pandas